Formatting: rename example to blinky and move function implementations from source to header
