+++ vehicle/examples/blinky/blinky.c
@@ -1,0 +1,18 @@
+/*
+This file is specifically meant for the STM32G441 dev board that was created and can be found in mkviii/hardware/example
+*/
+
+#include "blinky.h"
+
+int main(void) {
+  HAL_Init();
+  SystemClockConfig();
+  GpioInit();
+
+  while (1) {
+    HAL_Delay(100); // Fast
+    HAL_GPIO_TogglePin(PA1_GPIO_Port, PA1_Pin);
+  }
+  return 0;
+}
+
